+++ src/Model/User.php
@@ -5,7 +5,6 @@ use Illuminate\Auth\UserInterface;
 use Illuminate\Auth\Reminders\RemindableTrait;
 use Illuminate\Auth\Reminders\RemindableInterface;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Model as Eloquent;
 use Illuminate\Database\Eloquent\SoftDeletingTrait;
 use Illuminate\Support\Facades\Hash;
 use Orchestra\Notifier\NotifiableTrait;
@@ -36,6 +35,13 @@ class User extends Eloquent implements UserInterface, RemindableInterface, Recip
      * @var array
      */
     protected $hidden = array('password', 'remember_token');
+
+    /**
+     * The attributes that should be mutated to dates.
+     *
+     * @var array
+     */
+    protected $dates = array('deleted_at');
 
     /**
      * Has many and belongs to relationship with Role.