+++ src/Model/UserMeta.php
@@ -1,7 +1,6 @@
 <?php namespace Orchestra\Model;
 
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Model as Eloquent;
 
 class UserMeta extends Eloquent
 {