+++ docs/changes.md
@@ -5,11 +5,16 @@ title: Model Change Log
 
 ## Version 2.2 {#v2-2}
 
+### v2.2.1 {#v2-2-1}
+
+* Fixes missing `$model->isSoftDeleting()` method.
+
 ### v2.2.0 {#v2-2-0}
 
 * Bump minimum version to PHP v5.4.0.
 * Add `Orchestra\Model\User::notify()` method which utilize `Orchestra\Notifier\NotifiableTrait`.
 * Move duplicates code to utilize traits.
+* Integrate with `Illuminate\Database\Eloquent\SoftDeletingTrait`.
 
 ## Version 2.1 {#v2-1}